lxd/storage: Remove legacy dir implementation
